requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
aiohttp==3.9.1
gunicorn==21.2.0